WiP:acc2hipgpufort: Fix data clause translation
Minor changes:

* Log output tuning
* Log output for counter modification routines

+++ runtime/gpufortrt/test/fortran/test0002_structured_region_02.f90
@@ -1,6 +1,6 @@
 program main
   integer,parameter :: N = 10
-  integer(4) :: x(N),y(2*N)
+  integer :: x(N),y(2*N)
   !$acc init
   !$acc data copy(x,y(1:N),y(N+1:2*N))
   !$acc end data
